pyproject.toml
[tool.poetry]
name = "manga-ai-repairer-backend"
version = "0.1.0"
description = ""
authors = ["laoliu <axel10@163.com>"]
readme = "README.md"
packages = [{include = "manga_ai_repairer_backend"}]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
flask = "^3.0.0"
pyinstaller = "^6.3.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
